IID_Experiments_and_Strategies/requirements.txt
config==0.5.1
einops==0.4.1
matplotlib==3.5.1
numpy==1.21.2
pandas==1.4.1
Pillow==9.1.1
scikit_learn==1.1.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
